// frontend/src/lib/protocol/generated/TrafficUpdate.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PublishedTrafficTarget } from "./PublishedTrafficTarget";
import type { TrafficDelta } from "./TrafficDelta";

export type TrafficUpdate = { "type": "snapshot", "value": Array<PublishedTrafficTarget> } | { "type": "delta", "value": TrafficDelta };